+++ src/components/conversation/buttons/index.js
@@ -2,9 +2,9 @@ const React = require('react')
 const { Component } = React
 const styled = require('styled-components').default
 const Guid = require('guid')
-const route = require('../route')
+const route = require('../../../route')
 
-const Button = require('./button')
+const Button = require('../../button')
 
 const ButtonsContainer = styled.div`
   display: flex;
