+++ src/components/conversation/control-panel/zoom-button.js
@@ -1,7 +1,7 @@
 const React = require('react')
 const { Component } = React
 
-const Button = require('./button')
+const Button = require('../../button')
 
 class ZoomButton extends Component {
   render () {
