+++ src/components/conversation/control-panel/zoom-slider.js
@@ -2,7 +2,7 @@ const React = require('react')
 const { Component } = React
 const styled = require('styled-components').default
 
-const Button = require('./button')
+const Button = require('../../button')
 
 const PINK = '#f09'
 const BLACK = '#000'
