--- src/handle/edit-change.js
@@ -1,5 +1,0 @@
-module.exports = function () {
-  return function handleEditChange ({ editing }) {
-    this.setState({ editing })
-  }
-}
