--- src/handle/index.js
@@ -1,9 +1,0 @@
-module.exports = function ({ person }) {
-  return {
-    buttonDelete: require('./button-delete')(person),
-    editChange: require('./edit-change')(person),
-    linkAdd: require('./link-add')(person),
-    messageAdd: require('./message-add')(person),
-    messageDelete: require('./message-delete')(person),
-  }
-}
