--- src/handle/link-add.js
@@ -1,6 +1,0 @@
-module.exports = function ({ linkAdd }) {
-  return function handleLinkAdd (parentId, childId) {
-    linkAdd(parentId, childId)
-    this.setState(this.updateState(this.state))
-  }
-}
