--- src/handle/message-add.js
@@ -1,6 +1,0 @@
-module.exports = function ({ messageAdd }, cb) {
-  return function handleMessageAdd (nodeId) {
-    messageAdd(nodeId)
-    this.setState(this.updateState(this.state), cb)
-  }
-}
