--- src/handle/message-delete.js
@@ -1,7 +1,0 @@
-module.exports = function ({ messageDelete }) {
-  return function handleMessageDelete (nodeId, messageIndex) {
-    messageDelete(nodeId, messageIndex)
-    this.setState(this.updateState(this.state))
-  }
-}
-
